Create testing directory; Configure full stack
The current testing configuration has the following:

  * openvpn server
  * etcd server
  * coredns server
  * 3x openvpn clients

Docker has some odd DNS behaviors, so the DNS push settings can not be tested
wtih this setup, but it an effective stage for experimenting with the VPN
client/server and DNS server relationships.

I have manually confirmed that addresses added to the etcd KV store are
resolvable by VPN clients. Automating this test should not be too difficult.